Work on File repository
* refactor FileItem to be outside the
  repository class for clarity
