+++ src/AasxWpfControlLibrary/SelectFromRepositoryFlyout.xaml.cs
@@ -32,9 +32,9 @@ namespace AasxPackageExplorer
     {
         public event IFlyoutControlClosed ControlClosed;
 
-        public AasxFileRepository.FileItem ResultItem = null;
+        public AasxFileRepositoryItem ResultItem = null;
 
-        private List<AasxFileRepository.FileItem> _listFileItems;
+        private List<AasxFileRepositoryItem> _listFileItems;
 
         public SelectFromRepositoryFlyout()
         {
@@ -53,11 +53,11 @@ namespace AasxPackageExplorer
         {
         }
 
-        public bool LoadAasxRepoFile(IEnumerable<AasxFileRepository.FileItem> items = null)
+        public bool LoadAasxRepoFile(IEnumerable<AasxFileRepositoryItem> items = null)
         {
             try
             {
-                this._listFileItems = new List<AasxFileRepository.FileItem>();
+                this._listFileItems = new List<AasxFileRepositoryItem>();
 
                 if (items != null)
                 {
@@ -104,7 +104,7 @@ namespace AasxPackageExplorer
             var b = sender as Button;
             if (b?.Tag != null && this._listFileItems != null && this._listFileItems.Contains(b.Tag))
             {
-                this.ResultItem = b.Tag as AasxFileRepository.FileItem;
+                this.ResultItem = b.Tag as AasxFileRepositoryItem;
                 ControlClosed?.Invoke();
             }
         }
